ui: add shimmer-enhanced skeletons with Suspense for all public pages
- Add theme-colored shimmer sweep CSS animation using var(--primary)
- Upgrade Skeleton primitive with optional shimmer prop (default: true)
- Add domain-specific skeletons: HeroSection, FilterBar, BlogCard, EventCard, etc.
- Wrap all public data-fetching pages in explicit <Suspense>
- Extract data fetching into dedicated *-data components
- Remove redundant event/[slug]/loading.tsx (now inline fallback)
- Replace [...Array(N)] with explicit arrays to satisfy lint rules
- Respect prefers-reduced-motion for accessibility
