+++ app/event/list/event-list-data.tsx
@@ -1,0 +1,8 @@
+import { getCachedApprovedEvents } from '@/lib/server/events-public'
+import EventListClient from './event-list-client'
+
+export default async function EventListData() {
+  const initialEvents = await getCachedApprovedEvents()
+
+  return <EventListClient initialEvents={initialEvents || []} />
+}
