+++ app/project/list/page.tsx
@@ -1,74 +1,97 @@
-import { getTranslations } from 'next-intl/server'
-import { Footer } from '@/components/ui/footer'
-import { Navbar } from '@/components/ui/navbar'
-import { fetchProjectsWithSorting } from '@/lib/actions'
-import { getCategories } from '@/lib/categories'
-import { getCurrentUser } from '@/lib/server/auth'
-import { ProjectListClient } from './project-list-client'
+'use client'
+
+import { Suspense } from 'react'
+import { Skeleton } from '@/components/ui/skeleton'
+import ProjectListData from './project-list-data'
 
 type SearchParams = Promise<{ sort?: string; filter?: string }>
 
-export default async function ProjectListPage({ searchParams }: { searchParams: SearchParams }) {
-  const { sort, filter } = await searchParams
-  const t = await getTranslations('projectList')
-
-  const initialSort = sort === 'top' || sort === 'newest' ? sort : 'trending'
-  const initialFilter = filter || 'all'
-
-  const [currentUser, { projects: initialProjects }, categories] = await Promise.all([
-    getCurrentUser(),
-    fetchProjectsWithSorting(initialSort, initialFilter === 'all' ? undefined : initialFilter, 100),
-    getCategories(),
-  ])
-
-  const filterOptions = categories.map((cat) => ({
-    value: cat.name,
-    label: cat.display_name,
-  }))
-  const normalizedProjects = (initialProjects ?? []).map((project) => ({
-    ...project,
-    image: project.image ?? '/vibedev-guest-avatar.png',
-  }))
-
-  const user = currentUser
-    ? {
-        name: currentUser.name,
-        email: currentUser.email,
-        avatar: currentUser.avatar,
-        username: currentUser.username,
-        role: currentUser.role,
-      }
-    : null
-
+function ProjectListLoadingFallback() {
   return (
     <div className="min-h-screen bg-background">
       <div className="relative min-h-screen bg-grid-pattern">
         <div className="absolute inset-0 bg-gradient-to-b from-background/80 via-background/60 to-background/80"></div>
 
-        <Navbar
-          showNavigation={true}
-          isLoggedIn={!!currentUser}
-          user={user ?? undefined}
-        />
+        <nav className="h-16 w-full border-b bg-background/80 backdrop-blur-md" />
 
         <section className="relative bg-transparent py-12 pt-24">
           <div className="mx-auto max-w-7xl px-4 sm:px-6 lg:px-8">
+            {/* Page header */}
             <div className="mb-12 text-center">
-              <h1 className="mb-4 font-bold text-4xl text-foreground tracking-tight lg:text-5xl">{t('title')}</h1>
-              <p className="mx-auto max-w-2xl text-muted-foreground text-xl">{t('description')}</p>
+              <Skeleton className="mx-auto mb-4 h-10 w-64 md:h-12" />
+              <Skeleton className="mx-auto h-6 w-full max-w-2xl" />
             </div>
 
-            <ProjectListClient
-              initialProjects={normalizedProjects}
-              initialFilter={initialFilter}
-              initialSort={initialSort}
-              filterOptions={filterOptions}
-            />
+            {/* Filter bar */}
+            <div className="mb-8">
+              <div className="space-y-3 md:hidden">
+                <Skeleton className="mx-auto h-10 w-full max-w-sm rounded-full" />
+                <div className="grid grid-cols-2 gap-3">
+                  <Skeleton className="h-10 w-full rounded-lg" />
+                  <Skeleton className="h-10 w-full rounded-lg" />
+                </div>
+              </div>
+              <div className="hidden md:grid md:grid-cols-[1fr_auto_1fr] md:items-center">
+                <div className="justify-self-start">
+                  <Skeleton className="h-10 w-48 rounded-lg" />
+                </div>
+                <div className="justify-self-center">
+                  <Skeleton className="h-10 w-40 rounded-full" />
+                </div>
+                <div className="justify-self-end">
+                  <Skeleton className="h-10 w-44 rounded-lg" />
+                </div>
+              </div>
+            </div>
+
+            {/* Project grid */}
+            <div className="grid grid-cols-1 gap-6 md:grid-cols-2 lg:grid-cols-3">
+              {[1, 2, 3, 4, 5, 6, 7, 8, 9].map((i) => (
+                <div
+                  key={i}
+                  className="space-y-4"
+                >
+                  <div className="bg-muted relative overflow-hidden rounded-lg">
+                    <div className="aspect-video">
+                      <Skeleton className="h-full w-full" />
+                    </div>
+                  </div>
+                  <div className="space-y-2">
+                    <Skeleton className="h-5 w-3/4" />
+                    <Skeleton className="h-4 w-full" />
+                    <Skeleton className="h-4 w-2/3" />
+                    <div className="flex items-center gap-4">
+                      <div className="flex items-center gap-1">
+                        <Skeleton className="h-4 w-4" />
+                        <Skeleton className="h-4 w-4" />
+                      </div>
+                      <div className="flex items-center gap-1">
+                        <Skeleton className="h-4 w-4" />
+                        <Skeleton className="h-4 w-4" />
+                      </div>
+                    </div>
+                  </div>
+                </div>
+              ))}
+            </div>
           </div>
         </section>
 
-        <Footer />
+        <footer className="border-t bg-background py-12">
+          <div className="mx-auto max-w-7xl px-4">
+            <Skeleton className="mx-auto h-5 w-64" />
+            <Skeleton className="mx-auto mt-2 h-4 w-48" />
+          </div>
+        </footer>
       </div>
     </div>
+  )
+}
+
+export default function ProjectListPage({ searchParams }: { searchParams: SearchParams }) {
+  return (
+    <Suspense fallback={<ProjectListLoadingFallback />}>
+      <ProjectListData searchParams={searchParams} />
+    </Suspense>
   )
 }
